requirements.txt
aio-pika==8.0.3
aiofiles==0.8.0
aiohttp==3.8.1
aiormq==6.3.3
aiosignal==1.2.0
alembic==1.8.0
anyio==3.6.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
fake-useragent==0.1.11
fastapi==0.78.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Js2Py==0.71
Mako==1.2.1
MarkupSafe==2.1.1
multidict==6.0.2
outcome==1.2.0
pamqp==3.1.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.1
pyjsparser==2.7.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
python-multipart==0.0.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.2
starlette==0.19.1
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
uvicorn==0.18.2
wsproto==1.1.0
yarl==1.7.2
